requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.4
requests==2.32.1 
python-dotenv==1.0.1     
psycopg2-binary==2.9.9
pytest
pytest-cov